+++ CHANGELOG.md
@@ -15,12 +15,14 @@ See [CHANGELOG - 6to5](CHANGELOG-6to5.md) for the pre-4.0.0 version changelog.
 
 ## 6.0.14
 
+ * **Spec Compliancy**
+  * Update exponentiation operator precedence.
+  * Fix parser bug where arrow functions have a higher precedence than they should.
  * **Bug Fix**
   * Fix SystemJS module formatter exporting function parameters.
   * Ensure that invalid identifier JSX attribute keys are quoted when transforming to calls.
   * Fix ES3 property literal plugin.
   * Fix parameters after defaults in arrow functions refering to the wrong `arguments`.
-  * Fix parser bug where arrow functions have a higher precedence than they should.
 
 ## 6.0.13